refactor: rename usecases to use_cases
