+++ frontend/lib/injection_container.dart
@@ -3,12 +3,12 @@ import 'package:dio/dio.dart';
 import 'package:news_app_clean_architecture/features/daily_news/data/data_sources/remote/news_api_service.dart';
 import 'package:news_app_clean_architecture/features/daily_news/data/repository/article_repository_impl.dart';
 import 'package:news_app_clean_architecture/features/daily_news/domain/repository/article_repository.dart';
-import 'package:news_app_clean_architecture/features/daily_news/domain/usecases/get_article.dart';
+import 'package:news_app_clean_architecture/features/daily_news/domain/use_cases/get_article.dart';
 import 'package:news_app_clean_architecture/features/daily_news/presentation/bloc/article/remote/remote_article_bloc.dart';
 import 'features/daily_news/data/data_sources/local/app_database.dart';
-import 'features/daily_news/domain/usecases/get_saved_article.dart';
-import 'features/daily_news/domain/usecases/remove_article.dart';
-import 'features/daily_news/domain/usecases/save_article.dart';
+import 'features/daily_news/domain/use_cases/get_saved_article.dart';
+import 'features/daily_news/domain/use_cases/remove_article.dart';
+import 'features/daily_news/domain/use_cases/save_article.dart';
 import 'features/daily_news/presentation/bloc/article/local/local_article_bloc.dart';
 
 final sl = GetIt.instance;
